[PIPE-10] Moving test code to separate test module and fixtures to shared conftest.py file
